requirements.txt
playwright
feedparser
openai
python-dotenv
beautifulsoup4
